move vlaidator assingments into chain module
